requirements.txt
asgiref==3.8.1
blinker==1.9.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
cloudinary==1.44.1
colorama==0.4.6
crispy-bootstrap5==2025.6
dj-database-url==3.0.1
django-cloudinary-storage==0.3.0
django-crispy-forms
greenlet==3.2.2
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==25.0
pillow==11.3.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
python-dotenv==1.1.0
requests==2.32.4
six==1.17.0
SQLAlchemy==2.0.41
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.1.3
whitenoise
django-widget-tweaks
djangorestframework